Refactored so that entities with different responses can be used
The library `mysql-simple` seems to have a restriction that the response of the
`query` function must have the same type.. always. So, if I have two different
types of responses from the database e.g. different number of columns, I can't
use a generic query function. So I had to create a query wrapper per Entity that
I have. This is of course not ideal... but then I don't know how to do this
generically. For now, I'll go with this and hope to get wiser along the way.

Furthermore, the `Filter` data type has been split so that the entity to filter
on and the other conditions are kept separate. This is because I'll have cases
in which I want to filter on both e.g:

```
casefiles "sample" | documents "contract"
```

The documents need to be filtered on the case files from the previous step as
well as the title.