Updated the Entity data structure
Not every type of entity is a Maybe. For Example, if you have the following
entity:

```
CustomerEntity Nothing
```

it will represent the fact that `customers` table was searched and nothing was
found. If `CustomerEntity Nothing` is present in the history of the results,
then a new search for the Customers will not be made.

Basically, this change will help in fixing the following:

```
customers 1 | casefiles "sample" | customers *
```

Instead of returning all the customers that have a case file with name sample,
the history will be checked first and if `CustomerEntity Nothing` is found, then
no results will be returned.