Using a prefetched schema for preparing the queries
Without this evaluating pine expression would have to fetch the schema for
tables in each operation. Using a prefetched schema also helped with writing the
tests and keeps the functions pure i.e. we don't have to execute another query
and talk to the database to prepare the SQL corresponding the pine expression.