feat: support for schemas (table groups) when joining
Some points to keep in mind for future enhancements:

1. Schema is assumed to be public (if not specified) for the first operation

The following are the same:

```
user
public.user
```

2. The `schema` / table group, must be specified

```
tenant | requests.request -- schema is needed on the join if not `public`
```

3. Tests need to be fixed !!!

4. Primary key is always assumed to be `id`. Should be determined by looking at
the schema.

5. We are getting the table group (schema) information in the references which
isn't really needed now that I think about it. Instead we need the schema of the
table itself. I'll add that support when starting to look into auto-completion
and hinting.