Updated the syntax for specifying parent and child relations
This change avoids the need for backtracking i.e. the user doesn't have to go
back and change the expression. The modifier can be appended to the end of the
expression.

```
x | of: y
x | has: y
```
is now:
```
x | y :parent
x | y :child
```

This also means that the order of the modifiers doesn't matter anymore.